benchmark/requirements.txt
h5py
pytorch >= 1.0
tqdm
pandas